Fixing some conversion bugs in rabema with MinGW.
git-svn-id: http://svn.seqan.de/seqan/trunk@8129 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/rabema/build_gold_standard.h
@@ -260,7 +260,7 @@ size_t buildErrorCurvePoints(String<WeightedMatch> & errorCurve,
         // changes.
         if (foundWithTooLowScore) {
             if (beginPosition(finder) == hitBeginPosition) {
-                relativeScore = static_cast<int>(ceilAwayFromZero(100.0 * getScore(pattern) / length(read)));
+                relativeScore = static_cast<int>(ceilAwayFromZero(100.0 * static_cast<double>(getScore(pattern)) / static_cast<double>(length(read))));
                 appendValue(tempMatches, WeightedMatch(contigId, isForward, endPosition(finder) - 1, relativeScore, beginPosition(finder)));
                 TPosition currentBeginPosition = beginPosition(finder);
                 // Add the rest until we hit one with a different begin position.
